# packages/saas-platform-spec/admin-api.openapi.yaml
openapi: 3.1.0
...
info:
    title: SaaS-Plattform SuperAdmin-API
    version: 0.1.0-draft
    description: |
        Read- und Write-Operationen für die Plattform-Verwaltung:
        Tenants, Users, Subscriptions, Promo-Codes, Audit-Log,
        Manifest-Discovery, plus Plan-Catalog-Lookup.
        Authentifizierung via JWT mit Rolle SUPER_ADMIN; alle schreibenden
        Operationen verlangen zusätzlich einen TOTP-Code im Header
        X-Mfa-Code, sofern der Admin MFA aktiviert hat (sonst 401
        MFA_REQUIRED). Manifest ist nie Security-Quelle (siehe SPEC §10.4).
servers:
    - url: https://{host}/api/v1/admin
      variables:
          host:
              default: localhost:3000
              description: Konsumenten-spezifischer Host (Domain der jeweiligen App)
...
security:
    - bearerAuth: []
paths:
...
    /catalog/plans/{id}:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: Plan-Stamm Detail (inkl. Versionen)
            responses: { '200': { description: Detail } }
        patch:
            tags: [catalog]
            summary: Plan-Stamm aktualisieren (Label/Beschreibung/Icon/sortOrder)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogStampUpdate' } } }
            responses: { '200': { description: Aktualisiert } }
        delete:
            tags: [catalog]
            summary: Plan-Stamm soft-deleten
            x-mfa-required: true
            responses: { '200': { description: Gelöscht } }
    /catalog/plans/{id}/purge:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        delete:
            tags: [catalog]
            summary: Plan-Stamm hart löschen (nur ohne aktive Referenzen)
            x-mfa-required: true
            responses:
                '200': { description: Gepurged }
                '422': { description: Noch referenziert, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    /catalog/plans/{id}/versions:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: PlanVersionen eines Plans auflisten
            responses: { '200': { description: Liste } }
        post:
            tags: [catalog]
            summary: PlanVersion-Entwurf anlegen (unpublished, publishedAt=null)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogPlanVersionDraft' } } }
            responses: { '201': { description: Entwurf angelegt } }
...
    /catalog/promotions/{id}:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: Promotion Detail
            responses: { '200': { description: Detail } }
        patch:
            tags: [catalog]
            summary: Promotion aktualisieren
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogPromotionUpdate' } } }
            responses: { '200': { description: Aktualisiert } }
        delete:
            tags: [catalog]
            summary: Promotion soft-deleten
            x-mfa-required: true
            responses: { '200': { description: Gelöscht } }

    # ── Feature-/Quota-/Capability-Katalog (Discovery-Pflege) ──
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...
    schemas:
...
        CatalogPlanVersionDraft:
            type: object
            required: [features, quotas, monthlyNet, yearlyNet]
            properties:
                features:
                    type: array
                    items: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
                bundles:
                    type: array
                    items: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
                quotas:
                    type: object
                    additionalProperties: { type: number }
                    description: 'camelCase-Keys, z. B. {"members": 150}'
                monthlyNet:
                    { type: string, pattern: '^\d+(\.\d{1,2})?$', description: 'EUR-Decimal-String, nicht Cent' }
                yearlyNet: { type: string, pattern: '^\d+(\.\d{1,2})?$' }
                marketed: { type: boolean }
                changeNote: { type: string, maxLength: 2000 }
                baseVersionId: { type: string, nullable: true }
                validFrom: { type: string, nullable: true, description: 'YYYY-MM-DD oder ISO-8601' }
                validUntil: { type: string, nullable: true }
...
        CatalogStampUpdate:
            type: object
            description: 'Stamm-Update für Plan ODER BusinessType'
            properties:
                label: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, maxLength: 2000, nullable: true }
                icon: { type: string, maxLength: 64, nullable: true }
                sortOrder: { type: integer, minimum: 0, maximum: 10000 }
...
        CatalogPromotionUpdate:
            type: object
            description: 'Partial-Update einer Promotion (alle Felder optional)'
            properties:
                internalLabel: { type: string, maxLength: 120 }
                type: { type: string, enum: [percent, amount, intro, freeMonths] }
                value: { oneOf: [{ type: number }, { type: object }] }
                appliesTo: { type: array, items: { type: string } }
                targetType: { type: string, enum: [PLAN, BUNDLE, OFFER] }
                billingCycle: { type: string, enum: [monthly, yearly, both] }
                validFrom: { type: string }
                validTo: { type: string }
                priority: { type: integer }
                onlyLocales: { type: array, items: { type: string }, nullable: true }
                requiresCoupon: { type: boolean }
                codes: { type: array, items: { type: string } }
                color: { type: string, maxLength: 16 }
                i18n: { type: object, additionalProperties: true }

        # ────────── Common ──────────
...
        ErrorResponse:
            type: object
            required: [code, message]
            properties:
                code:
                    type: string
                    enum:
                        - LIMIT_EXCEEDED
                        - FEATURE_NOT_INCLUDED
                        - MFA_REQUIRED
                        - PROMO_NOT_FOUND
                        - PROMO_EXPIRED
                        - PROMO_EXHAUSTED
                        - PROMO_NOT_APPLICABLE
                        - PROMO_FIRST_TIME_ONLY
                        - PROMO_ZERO_INVOICE_BLOCKED
                        - VALIDATION_FAILED
                        - FORBIDDEN
                        - NOT_FOUND
                message: { type: string }
                details: { type: object, additionalProperties: true }

        # ────────── Plan-Catalog (mirror of plans.yaml) ──────────
...